% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salReturn.R
\name{salReturn_insert}
\alias{salReturn_insert}
\title{更新插入list表和表头表体数据}
\usage{
salReturn_insert(dms_token)
}
\arguments{
\item{dms_token}{第二个参数}
}
\value{
两个数的和
}
\description{
更新插入list表和表头表体数据
}
\examples{
salReturn_insert
}
